Refactor to platform-level flag protection
Reduce duplication following the general advice around this topic that
it's ultimately up to apps how they do flags, and going by platform
conventions seems like the most straightforward solution from a library
perspective.
