+++ src/internal/win.js
@@ -93,24 +93,23 @@ export function getQuoteFunction(shellName) {
 }
 
 /**
+ * Remove any prefix from the provided argument that might be interpreted as a
+ * flag on Windows systems for CMD.
+ *
+ * @param {string} arg The argument to update.
+ * @returns {string} The updated argument.
+ */
+function stripFlagPrefix(arg) {
+  return arg.replace(/^(?:-+|\/+)/gu, "");
+}
+
+/**
  * Returns a function to protect against flag injection.
  *
- * @param {string | symbol} shellName The name of a Windows shell.
- * @returns {Function | undefined} A function to protect against flag injection.
+ * @returns {Function} A function to protect against flag injection.
  */
-export function getFlagProtectionFunction(shellName) {
-  if (shellName === noShell) {
-    return nosh.getFlagProtectionFunction();
-  }
-
-  switch (shellName.toLowerCase()) {
-    case binCmd: {
-      return cmd.getFlagProtectionFunction();
-    }
-    case binPowerShell: {
-      return powershell.getFlagProtectionFunction();
-    }
-  }
+export function getFlagProtectionFunction() {
+  return stripFlagPrefix;
 }
 
 /**
