+++ src/internal/win/no-shell.js
@@ -47,23 +47,3 @@ function unsupported() {
 export function getQuoteFunction() {
   return [unsupported, unsupported];
 }
-
-/**
- * Remove any prefix from the provided argument that might be interpreted as a
- * flag on Windows systems.
- *
- * @param {string} arg The argument to update.
- * @returns {string} The updated argument.
- */
-function stripFlagPrefix(arg) {
-  return arg.replace(/^(?:-+|\/+)/gu, "");
-}
-
-/**
- * Returns a function to protect against flag injection for Windows systems.
- *
- * @returns {Function} A function to protect against flag injection.
- */
-export function getFlagProtectionFunction() {
-  return stripFlagPrefix;
-}
