+++ test/fixtures/win.js
@@ -3849,408 +3849,138 @@ export const escape = {
 };
 
 export const flag = {
-  [null]: {
-    "sample strings": [
-      {
-        input: "foobar",
-        expected: { unquoted: "foobar", quoted: "foobar" },
-      },
-    ],
-    "single hyphen (-)": [
-      {
-        input: "-a",
-        expected: { unquoted: "a", quoted: "a" },
-      },
-      {
-        input: "a-",
-        expected: { unquoted: "a-", quoted: "a-" },
-      },
-      {
-        input: "-a-",
-        expected: { unquoted: "a-", quoted: "a-" },
-      },
-      {
-        input: "-ab",
-        expected: { unquoted: "ab", quoted: "ab" },
-      },
-      {
-        input: "a-b",
-        expected: { unquoted: "a-b", quoted: "a-b" },
-      },
-      {
-        input: "-a-b",
-        expected: { unquoted: "a-b", quoted: "a-b" },
-      },
-      {
-        input: "-a=b",
-        expected: { unquoted: "a=b", quoted: "a=b" },
-      },
-    ],
-    "double hyphen (--)": [
-      {
-        input: "--a",
-        expected: { unquoted: "a", quoted: "a" },
-      },
-      {
-        input: "a--",
-        expected: { unquoted: "a--", quoted: "a--" },
-      },
-      {
-        input: "--a--",
-        expected: { unquoted: "a--", quoted: "a--" },
-      },
-      {
-        input: "--ab",
-        expected: { unquoted: "ab", quoted: "ab" },
-      },
-      {
-        input: "a--b",
-        expected: { unquoted: "a--b", quoted: "a--b" },
-      },
-      {
-        input: "--a--b",
-        expected: { unquoted: "a--b", quoted: "a--b" },
-      },
-      {
-        input: "--a=b",
-        expected: { unquoted: "a=b", quoted: "a=b" },
-      },
-    ],
-    "many hyphens (/-{3,}/)": [
-      {
-        input: "---a",
-        expected: { unquoted: "a", quoted: "a" },
-      },
-      {
-        input: "---ab",
-        expected: { unquoted: "ab", quoted: "ab" },
-      },
-      {
-        input: "---a=b",
-        expected: { unquoted: "a=b", quoted: "a=b" },
-      },
-    ],
-    "forward slash (/)": [
-      {
-        input: "/a",
-        expected: { unquoted: "a", quoted: "a" },
-      },
-      {
-        input: "a/",
-        expected: { unquoted: "a/", quoted: "a/" },
-      },
-      {
-        input: "/a/",
-        expected: { unquoted: "a/", quoted: "a/" },
-      },
-      {
-        input: "/ab",
-        expected: { unquoted: "ab", quoted: "ab" },
-      },
-      {
-        input: "a/b",
-        expected: { unquoted: "a/b", quoted: "a/b" },
-      },
-      {
-        input: "/a/b",
-        expected: { unquoted: "a/b", quoted: "a/b" },
-      },
-    ],
-    "multiple forward slashes (//{2,}/)": [
-      {
-        input: "//a",
-        expected: { unquoted: "a", quoted: "a" },
-      },
-      {
-        input: "a//",
-        expected: { unquoted: "a//", quoted: "a//" },
-      },
-      {
-        input: "//a//",
-        expected: { unquoted: "a//", quoted: "a//" },
-      },
-      {
-        input: "//ab",
-        expected: { unquoted: "ab", quoted: "ab" },
-      },
-      {
-        input: "a//b",
-        expected: { unquoted: "a//b", quoted: "a//b" },
-      },
-      {
-        input: "//a//b",
-        expected: { unquoted: "a//b", quoted: "a//b" },
-      },
-    ],
-  },
-  [binCmd]: {
-    "sample strings": [
-      {
-        input: "foobar",
-        expected: { unquoted: "foobar", quoted: '"foobar"' },
-      },
-    ],
-    "single hyphen (-)": [
-      {
-        input: "-a",
-        expected: { unquoted: "a", quoted: '"a"' },
-      },
-      {
-        input: "a-",
-        expected: { unquoted: "a-", quoted: '"a-"' },
-      },
-      {
-        input: "-a-",
-        expected: { unquoted: "a-", quoted: '"a-"' },
-      },
-      {
-        input: "-ab",
-        expected: { unquoted: "ab", quoted: '"ab"' },
-      },
-      {
-        input: "a-b",
-        expected: { unquoted: "a-b", quoted: '"a-b"' },
-      },
-      {
-        input: "-a-b",
-        expected: { unquoted: "a-b", quoted: '"a-b"' },
-      },
-      {
-        input: "-a=b",
-        expected: { unquoted: "a=b", quoted: '"a=b"' },
-      },
-    ],
-    "double hyphen (--)": [
-      {
-        input: "--a",
-        expected: { unquoted: "a", quoted: '"a"' },
-      },
-      {
-        input: "a--",
-        expected: { unquoted: "a--", quoted: '"a--"' },
-      },
-      {
-        input: "--a--",
-        expected: { unquoted: "a--", quoted: '"a--"' },
-      },
-      {
-        input: "--ab",
-        expected: { unquoted: "ab", quoted: '"ab"' },
-      },
-      {
-        input: "a--b",
-        expected: { unquoted: "a--b", quoted: '"a--b"' },
-      },
-      {
-        input: "--a--b",
-        expected: { unquoted: "a--b", quoted: '"a--b"' },
-      },
-      {
-        input: "--a=b",
-        expected: { unquoted: "a=b", quoted: '"a=b"' },
-      },
-    ],
-    "many hyphens (/-{3,}/)": [
-      {
-        input: "---a",
-        expected: { unquoted: "a", quoted: '"a"' },
-      },
-      {
-        input: "---ab",
-        expected: { unquoted: "ab", quoted: '"ab"' },
-      },
-      {
-        input: "---a=b",
-        expected: { unquoted: "a=b", quoted: '"a=b"' },
-      },
-    ],
-    "forward slash (/)": [
-      {
-        input: "/a",
-        expected: { unquoted: "a", quoted: '"a"' },
-      },
-      {
-        input: "a/",
-        expected: { unquoted: "a/", quoted: '"a/"' },
-      },
-      {
-        input: "/a/",
-        expected: { unquoted: "a/", quoted: '"a/"' },
-      },
-      {
-        input: "/ab",
-        expected: { unquoted: "ab", quoted: '"ab"' },
-      },
-      {
-        input: "a/b",
-        expected: { unquoted: "a/b", quoted: '"a/b"' },
-      },
-      {
-        input: "/a/b",
-        expected: { unquoted: "a/b", quoted: '"a/b"' },
-      },
-    ],
-    "multiple forward slashes (//{2,}/)": [
-      {
-        input: "//a",
-        expected: { unquoted: "a", quoted: '"a"' },
-      },
-      {
-        input: "a//",
-        expected: { unquoted: "a//", quoted: '"a//"' },
-      },
-      {
-        input: "//a//",
-        expected: { unquoted: "a//", quoted: '"a//"' },
-      },
-      {
-        input: "//ab",
-        expected: { unquoted: "ab", quoted: '"ab"' },
-      },
-      {
-        input: "a//b",
-        expected: { unquoted: "a//b", quoted: '"a//b"' },
-      },
-      {
-        input: "//a//b",
-        expected: { unquoted: "a//b", quoted: '"a//b"' },
-      },
-    ],
-  },
-  [binPowerShell]: {
-    "sample strings": [
-      {
-        input: "foobar",
-        expected: { unquoted: "foobar", quoted: "'foobar'" },
-      },
-    ],
-    "single hyphen (-)": [
-      {
-        input: "-a",
-        expected: { unquoted: "a", quoted: "'a'" },
-      },
-      {
-        input: "a-",
-        expected: { unquoted: "a-", quoted: "'a-'" },
-      },
-      {
-        input: "-a-",
-        expected: { unquoted: "a-", quoted: "'a-'" },
-      },
-      {
-        input: "-ab",
-        expected: { unquoted: "ab", quoted: "'ab'" },
-      },
-      {
-        input: "a-b",
-        expected: { unquoted: "a-b", quoted: "'a-b'" },
-      },
-      {
-        input: "-a-b",
-        expected: { unquoted: "a-b", quoted: "'a-b'" },
-      },
-      {
-        input: "-a=b",
-        expected: { unquoted: "a=b", quoted: "'a=b'" },
-      },
-    ],
-    "double hyphen (--)": [
-      {
-        input: "--a",
-        expected: { unquoted: "a", quoted: "'a'" },
-      },
-      {
-        input: "a--",
-        expected: { unquoted: "a--", quoted: "'a--'" },
-      },
-      {
-        input: "--a--",
-        expected: { unquoted: "a--", quoted: "'a--'" },
-      },
-      {
-        input: "--ab",
-        expected: { unquoted: "ab", quoted: "'ab'" },
-      },
-      {
-        input: "a--b",
-        expected: { unquoted: "a--b", quoted: "'a--b'" },
-      },
-      {
-        input: "--a--b",
-        expected: { unquoted: "a--b", quoted: "'a--b'" },
-      },
-      {
-        input: "--a=b",
-        expected: { unquoted: "a=b", quoted: "'a=b'" },
-      },
-    ],
-    "many hyphens (/-{3,}/)": [
-      {
-        input: "---a",
-        expected: { unquoted: "a", quoted: "'a'" },
-      },
-      {
-        input: "---ab",
-        expected: { unquoted: "ab", quoted: "'ab'" },
-      },
-      {
-        input: "---a=b",
-        expected: { unquoted: "a=b", quoted: "'a=b'" },
-      },
-    ],
-    "forward slash (/)": [
-      {
-        input: "/a",
-        expected: { unquoted: "a", quoted: "'a'" },
-      },
-      {
-        input: "a/",
-        expected: { unquoted: "a/", quoted: "'a/'" },
-      },
-      {
-        input: "/a/",
-        expected: { unquoted: "a/", quoted: "'a/'" },
-      },
-      {
-        input: "/ab",
-        expected: { unquoted: "ab", quoted: "'ab'" },
-      },
-      {
-        input: "a/b",
-        expected: { unquoted: "a/b", quoted: "'a/b'" },
-      },
-      {
-        input: "/a/b",
-        expected: { unquoted: "a/b", quoted: "'a/b'" },
-      },
-    ],
-    "multiple forward slashes (//{2,}/)": [
-      {
-        input: "//a",
-        expected: { unquoted: "a", quoted: "'a'" },
-      },
-      {
-        input: "a//",
-        expected: { unquoted: "a//", quoted: "'a//'" },
-      },
-      {
-        input: "//a//",
-        expected: { unquoted: "a//", quoted: "'a//'" },
-      },
-      {
-        input: "//ab",
-        expected: { unquoted: "ab", quoted: "'ab'" },
-      },
-      {
-        input: "a//b",
-        expected: { unquoted: "a//b", quoted: "'a//b'" },
-      },
-      {
-        input: "//a//b",
-        expected: { unquoted: "a//b", quoted: "'a//b'" },
-      },
-    ],
-  },
+  "sample strings": [
+    {
+      input: "foobar",
+      expected: { unquoted: "foobar", quoted: "foobar" },
+    },
+  ],
+  "single hyphen (-)": [
+    {
+      input: "-a",
+      expected: { unquoted: "a", quoted: "a" },
+    },
+    {
+      input: "a-",
+      expected: { unquoted: "a-", quoted: "a-" },
+    },
+    {
+      input: "-a-",
+      expected: { unquoted: "a-", quoted: "a-" },
+    },
+    {
+      input: "-ab",
+      expected: { unquoted: "ab", quoted: "ab" },
+    },
+    {
+      input: "a-b",
+      expected: { unquoted: "a-b", quoted: "a-b" },
+    },
+    {
+      input: "-a-b",
+      expected: { unquoted: "a-b", quoted: "a-b" },
+    },
+    {
+      input: "-a=b",
+      expected: { unquoted: "a=b", quoted: "a=b" },
+    },
+  ],
+  "double hyphen (--)": [
+    {
+      input: "--a",
+      expected: { unquoted: "a", quoted: "a" },
+    },
+    {
+      input: "a--",
+      expected: { unquoted: "a--", quoted: "a--" },
+    },
+    {
+      input: "--a--",
+      expected: { unquoted: "a--", quoted: "a--" },
+    },
+    {
+      input: "--ab",
+      expected: { unquoted: "ab", quoted: "ab" },
+    },
+    {
+      input: "a--b",
+      expected: { unquoted: "a--b", quoted: "a--b" },
+    },
+    {
+      input: "--a--b",
+      expected: { unquoted: "a--b", quoted: "a--b" },
+    },
+    {
+      input: "--a=b",
+      expected: { unquoted: "a=b", quoted: "a=b" },
+    },
+  ],
+  "many hyphens (/-{3,}/)": [
+    {
+      input: "---a",
+      expected: { unquoted: "a", quoted: "a" },
+    },
+    {
+      input: "---ab",
+      expected: { unquoted: "ab", quoted: "ab" },
+    },
+    {
+      input: "---a=b",
+      expected: { unquoted: "a=b", quoted: "a=b" },
+    },
+  ],
+  "forward slash (/)": [
+    {
+      input: "/a",
+      expected: { unquoted: "a", quoted: "a" },
+    },
+    {
+      input: "a/",
+      expected: { unquoted: "a/", quoted: "a/" },
+    },
+    {
+      input: "/a/",
+      expected: { unquoted: "a/", quoted: "a/" },
+    },
+    {
+      input: "/ab",
+      expected: { unquoted: "ab", quoted: "ab" },
+    },
+    {
+      input: "a/b",
+      expected: { unquoted: "a/b", quoted: "a/b" },
+    },
+    {
+      input: "/a/b",
+      expected: { unquoted: "a/b", quoted: "a/b" },
+    },
+  ],
+  "multiple forward slashes (//{2,}/)": [
+    {
+      input: "//a",
+      expected: { unquoted: "a", quoted: "a" },
+    },
+    {
+      input: "a//",
+      expected: { unquoted: "a//", quoted: "a//" },
+    },
+    {
+      input: "//a//",
+      expected: { unquoted: "a//", quoted: "a//" },
+    },
+    {
+      input: "//ab",
+      expected: { unquoted: "ab", quoted: "ab" },
+    },
+    {
+      input: "a//b",
+      expected: { unquoted: "a//b", quoted: "a//b" },
+    },
+    {
+      input: "//a//b",
+      expected: { unquoted: "a//b", quoted: "a//b" },
+    },
+  ],
 };
 
 export const quote = {
