+++ test/unit/unix/shells.test.js
@@ -27,7 +27,6 @@ const shells = {
 
 for (const [shellName, shellExports] of Object.entries(shells)) {
   const escapeFixtures = Object.values(fixtures.escape[shellName]).flat();
-  const flagFixtures = Object.values(fixtures.flag[shellName]).flat();
   const quoteFixtures = Object.values(fixtures.quote[shellName]).flat();
   const redosFixtures = fixtures.redos();
 
@@ -60,31 +59,6 @@ for (const [shellName, shellExports] of Object.entries(shells)) {
       t.pass();
     });
   }
-
-  for (const { input, expected } of flagFixtures) {
-    test(macros.flag, {
-      expected: expected.unquoted,
-      input,
-      getFlagProtectionFunction: shellExports.getFlagProtectionFunction,
-      shellName,
-    });
-  }
-
-  testProp(
-    `flag protection function for ${shellName}`,
-    [fc.string()],
-    (t, arg) => {
-      const flagProtect = shellExports.getFlagProtectionFunction();
-      const result = flagProtect(arg);
-      t.is(typeof result, "string");
-    },
-  );
-
-  test(`flag protection performance for ${shellName}`, macros.duration, {
-    arbitraries: [fc.string({ size: "xlarge" })],
-    maxMillis: 50,
-    setup: shellExports.getFlagProtectionFunction,
-  });
 
   if (shellExports !== nosh) {
     for (const { input, expected } of quoteFixtures) {
